requirements.txt
cryptography==41.0.3
PyJWT==2.8.0
pymongo==4.5.0 